fix: keyboard navigation breaking
Fixes the keyboard navigation breaking in certain scenarios.

The issue was that `window.getSelection()` would not always return the currently focused element when checking if we need to disable our custom keybindings. So e.g. when clicking the search input and then clicking a file in the file list, the focused element would still be the search input, leading to our keybindings not working.